requirements.txt
asgiref==3.3.1
black==23.3.0
certifi==2020.12.5
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.3
Django==3.1.6
et-xmlfile==1.0.1
idna==2.10
jdcal==1.4.1
Jinja2==2.11.3
MarkupSafe==1.1.1
mypy-extensions==1.0.0
numpy==1.21.4
openpyxl==3.0.6
packaging==23.1
pandas==1.2.2
pathspec==0.11.1
platformdirs==3.5.1
psycopg2-binary==2.8.6
PyMySQL==1.0.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tomli==2.0.1
typing_extensions==4.6.3
tzdata==2023.4
urllib3==1.26.4
XlsxWriter==1.3.7